Introduced getActivityInstance and getIdlingResourceActivityInstance functions and removed code duplication
