+++ app/src/androidTest/java/com/alex/mapnotes/idlingresources/ViewTextIdlingResource.kt
@@ -1,12 +1,10 @@
 package com.alex.mapnotes.idlingresources
 
-import android.app.Activity
 import android.widget.TextView
 import androidx.annotation.IdRes
 import androidx.annotation.StringRes
 import androidx.test.espresso.IdlingResource
-import androidx.test.runner.lifecycle.ActivityLifecycleMonitorRegistry
-import androidx.test.runner.lifecycle.Stage
+import com.alex.mapnotes.getIdlingResourceActivityInstance
 
 class ViewTextIdlingResource(
     @IdRes private val expectedViewId: Int,
@@ -20,8 +18,8 @@ class ViewTextIdlingResource(
     }
 
     override fun isIdleNow(): Boolean {
-        val view: TextView? = getActivityInstance().findViewById(expectedViewId)
-        val expectedText = getActivityInstance().getString(expectedViewText)
+        val view: TextView? = getIdlingResourceActivityInstance().findViewById(expectedViewId)
+        val expectedText = getIdlingResourceActivityInstance().getString(expectedViewText)
         val isIdle = if (view != null) {
             view.text.toString() == expectedText
         } else {
@@ -36,14 +34,5 @@ class ViewTextIdlingResource(
 
     override fun registerIdleTransitionCallback(callback: IdlingResource.ResourceCallback?) {
         resourceCallback = callback
-    }
-
-    private fun getActivityInstance(): Activity {
-        var currentActivity: Activity? = null
-        val resumedActivities = ActivityLifecycleMonitorRegistry.getInstance().getActivitiesInStage(Stage.RESUMED)
-        if (resumedActivities.iterator().hasNext()) {
-            currentActivity = resumedActivities.iterator().next()
-        }
-        return currentActivity!!
     }
 }
